+++ src/main/java/com/ametrinstudios/ametrin/Ametrin.java
@@ -1,7 +1,6 @@
 package com.ametrinstudios.ametrin;
 
 import com.ametrinstudios.ametrin.util.VanillaCompat;
-import com.ametrinstudios.ametrin.world.AmetrinEntityTypes;
 import net.neoforged.bus.api.IEventBus;
 import net.neoforged.fml.common.Mod;
 import net.neoforged.fml.event.lifecycle.FMLCommonSetupEvent;
@@ -13,7 +12,6 @@ public final class Ametrin {
     public static final String MOD_ID = "ametrin";
 
     public Ametrin(IEventBus modBus) {
-        AmetrinEntityTypes.REGISTER.register(modBus);
         modBus.addListener(Ametrin::commonSetup);
     }