+++ src/main/java/com/ametrinstudios/ametrin/world/gen/feature/tree/CustomTreeFeature.java
@@ -17,7 +17,7 @@ import java.util.function.BiConsumer;
 /**
  * This is an alternative method of adding trees, if possible use the vanilla way!
  */
-public abstract class CustomTreeFeature extends TreeFeature{
+public abstract class CustomTreeFeature extends TreeFeature {
     public CustomTreeFeature(Codec<TreeConfiguration> codec) {
         super(codec);
     }